refactor: separate engine to new process service
This includes moving other code which is generic conceptually even if not in implementation; changes to this code will be made later.
